+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add Compodoc icon [Thanks dotpresso](https://github.com/BeardedBear/bearded-icons/pull/46)
 - Add Julia icon [Issue #24](https://github.com/BeardedBear/bearded-icons/issues/24)
 - Add CSV icon [Issue #25](https://github.com/BeardedBear/bearded-icons/issues/25)
+- Add Karma icon [Issue #26](https://github.com/BeardedBear/bearded-icons/issues/26)
 
 ## 0.14.0 - 2021-07-24