+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add Karma icon [Issue #26](https://github.com/BeardedBear/bearded-icons/issues/26)
 - Add tsconfig icon [Issue #27](https://github.com/BeardedBear/bearded-icons/issues/27)
 - Add Travis CI icon [Issue #28](https://github.com/BeardedBear/bearded-icons/issues/28)
+- Add Razor and cshtml icons [Issue #30](https://github.com/BeardedBear/bearded-icons/issues/30)
 
 ## 0.14.0 - 2021-07-24